Change Vec file structure
